% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedSeasonStat.r
\name{AdvancedSeasonStat}
\alias{AdvancedSeasonStat}
\title{AdvancedSeasonStat Class}
\description{
AdvancedSeasonStat Class

AdvancedSeasonStat Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{season}}{}

\item{\code{team}}{}

\item{\code{conference}}{}

\item{\code{offense}}{}

\item{\code{defense}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdvancedSeasonStat$new()}}
\item \href{#method-toJSON}{\code{AdvancedSeasonStat$toJSON()}}
\item \href{#method-fromJSON}{\code{AdvancedSeasonStat$fromJSON()}}
\item \href{#method-toJSONString}{\code{AdvancedSeasonStat$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AdvancedSeasonStat$fromJSONString()}}
\item \href{#method-clone}{\code{AdvancedSeasonStat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStat$new(season, team, conference, offense, defense)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStat$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStat$fromJSON(AdvancedSeasonStatJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStat$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStat$fromJSONString(AdvancedSeasonStatJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
