% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedSeasonStatOffense.r
\name{AdvancedSeasonStatOffense}
\alias{AdvancedSeasonStatOffense}
\title{AdvancedSeasonStatOffense Class}
\description{
AdvancedSeasonStatOffense Class

AdvancedSeasonStatOffense Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{plays}}{}

\item{\code{drives}}{}

\item{\code{ppa}}{}

\item{\code{totalPPA}}{}

\item{\code{successRate}}{}

\item{\code{explosiveness}}{}

\item{\code{powerSuccess}}{}

\item{\code{stuffRate}}{}

\item{\code{lineYards}}{}

\item{\code{lineYardsTotal}}{}

\item{\code{secondLevelYards}}{}

\item{\code{secondLevelYardsTotal}}{}

\item{\code{openFieldYards}}{}

\item{\code{openFieldYardsTotal}}{}

\item{\code{pointsPerOpportunity}}{}

\item{\code{fieldPosition}}{}

\item{\code{havoc}}{}

\item{\code{standardDowns}}{}

\item{\code{passingDowns}}{}

\item{\code{rushingPlays}}{}

\item{\code{passingPlays}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdvancedSeasonStatOffense$new()}}
\item \href{#method-toJSON}{\code{AdvancedSeasonStatOffense$toJSON()}}
\item \href{#method-fromJSON}{\code{AdvancedSeasonStatOffense$fromJSON()}}
\item \href{#method-toJSONString}{\code{AdvancedSeasonStatOffense$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AdvancedSeasonStatOffense$fromJSONString()}}
\item \href{#method-clone}{\code{AdvancedSeasonStatOffense$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffense$new(
  plays,
  drives,
  ppa,
  totalPPA,
  successRate,
  explosiveness,
  powerSuccess,
  stuffRate,
  lineYards,
  lineYardsTotal,
  secondLevelYards,
  secondLevelYardsTotal,
  openFieldYards,
  openFieldYardsTotal,
  pointsPerOpportunity,
  fieldPosition,
  havoc,
  standardDowns,
  passingDowns,
  rushingPlays,
  passingPlays
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffense$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffense$fromJSON(AdvancedSeasonStatOffenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffense$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffense$fromJSONString(AdvancedSeasonStatOffenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffense$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
