% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedSeasonStatOffenseFieldPosition.r
\name{AdvancedSeasonStatOffenseFieldPosition}
\alias{AdvancedSeasonStatOffenseFieldPosition}
\title{AdvancedSeasonStatOffenseFieldPosition Class}
\description{
AdvancedSeasonStatOffenseFieldPosition Class

AdvancedSeasonStatOffenseFieldPosition Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{averageStart}}{}

\item{\code{averagePredictedPoints}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdvancedSeasonStatOffenseFieldPosition$new()}}
\item \href{#method-toJSON}{\code{AdvancedSeasonStatOffenseFieldPosition$toJSON()}}
\item \href{#method-fromJSON}{\code{AdvancedSeasonStatOffenseFieldPosition$fromJSON()}}
\item \href{#method-toJSONString}{\code{AdvancedSeasonStatOffenseFieldPosition$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AdvancedSeasonStatOffenseFieldPosition$fromJSONString()}}
\item \href{#method-clone}{\code{AdvancedSeasonStatOffenseFieldPosition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseFieldPosition$new(
  averageStart,
  averagePredictedPoints
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseFieldPosition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseFieldPosition$fromJSON(
  AdvancedSeasonStatOffenseFieldPositionJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseFieldPosition$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseFieldPosition$fromJSONString(
  AdvancedSeasonStatOffenseFieldPositionJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseFieldPosition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
