% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedSeasonStatOffenseRushingPlays.r
\name{AdvancedSeasonStatOffenseRushingPlays}
\alias{AdvancedSeasonStatOffenseRushingPlays}
\title{AdvancedSeasonStatOffenseRushingPlays Class}
\description{
AdvancedSeasonStatOffenseRushingPlays Class

AdvancedSeasonStatOffenseRushingPlays Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rate}}{}

\item{\code{ppa}}{}

\item{\code{totalPPA}}{}

\item{\code{successRate}}{}

\item{\code{explosiveness}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdvancedSeasonStatOffenseRushingPlays$new()}}
\item \href{#method-toJSON}{\code{AdvancedSeasonStatOffenseRushingPlays$toJSON()}}
\item \href{#method-fromJSON}{\code{AdvancedSeasonStatOffenseRushingPlays$fromJSON()}}
\item \href{#method-toJSONString}{\code{AdvancedSeasonStatOffenseRushingPlays$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AdvancedSeasonStatOffenseRushingPlays$fromJSONString()}}
\item \href{#method-clone}{\code{AdvancedSeasonStatOffenseRushingPlays$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseRushingPlays$new(
  rate,
  ppa,
  totalPPA,
  successRate,
  explosiveness
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseRushingPlays$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseRushingPlays$fromJSON(
  AdvancedSeasonStatOffenseRushingPlaysJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseRushingPlays$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseRushingPlays$fromJSONString(
  AdvancedSeasonStatOffenseRushingPlaysJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseRushingPlays$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
