% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedSeasonStatOffenseStandardDowns.r
\name{AdvancedSeasonStatOffenseStandardDowns}
\alias{AdvancedSeasonStatOffenseStandardDowns}
\title{AdvancedSeasonStatOffenseStandardDowns Class}
\description{
AdvancedSeasonStatOffenseStandardDowns Class

AdvancedSeasonStatOffenseStandardDowns Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rate}}{}

\item{\code{ppa}}{}

\item{\code{successRate}}{}

\item{\code{explosiveness}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdvancedSeasonStatOffenseStandardDowns$new()}}
\item \href{#method-toJSON}{\code{AdvancedSeasonStatOffenseStandardDowns$toJSON()}}
\item \href{#method-fromJSON}{\code{AdvancedSeasonStatOffenseStandardDowns$fromJSON()}}
\item \href{#method-toJSONString}{\code{AdvancedSeasonStatOffenseStandardDowns$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AdvancedSeasonStatOffenseStandardDowns$fromJSONString()}}
\item \href{#method-clone}{\code{AdvancedSeasonStatOffenseStandardDowns$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseStandardDowns$new(
  rate,
  ppa,
  successRate,
  explosiveness
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseStandardDowns$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseStandardDowns$fromJSON(
  AdvancedSeasonStatOffenseStandardDownsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseStandardDowns$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseStandardDowns$fromJSONString(
  AdvancedSeasonStatOffenseStandardDownsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdvancedSeasonStatOffenseStandardDowns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
