% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScorePlayersAverage.r
\name{BoxScorePlayersAverage}
\alias{BoxScorePlayersAverage}
\title{BoxScorePlayersAverage Class}
\description{
BoxScorePlayersAverage Class

BoxScorePlayersAverage Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{total}}{}

\item{\code{quarter1}}{}

\item{\code{quarter2}}{}

\item{\code{quarter3}}{}

\item{\code{quarter4}}{}

\item{\code{rushing}}{}

\item{\code{passing}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScorePlayersAverage$new()}}
\item \href{#method-toJSON}{\code{BoxScorePlayersAverage$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScorePlayersAverage$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScorePlayersAverage$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScorePlayersAverage$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScorePlayersAverage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersAverage$new(
  total,
  quarter1,
  quarter2,
  quarter3,
  quarter4,
  rushing,
  passing
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersAverage$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersAverage$fromJSON(BoxScorePlayersAverageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersAverage$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersAverage$fromJSONString(BoxScorePlayersAverageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersAverage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
