% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScorePlayersUsage.r
\name{BoxScorePlayersUsage}
\alias{BoxScorePlayersUsage}
\title{BoxScorePlayersUsage Class}
\description{
BoxScorePlayersUsage Class

BoxScorePlayersUsage Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{player}}{}

\item{\code{team}}{}

\item{\code{position}}{}

\item{\code{total}}{}

\item{\code{quarter1}}{}

\item{\code{quarter2}}{}

\item{\code{quarter3}}{}

\item{\code{quarter4}}{}

\item{\code{rushing}}{}

\item{\code{passing}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScorePlayersUsage$new()}}
\item \href{#method-toJSON}{\code{BoxScorePlayersUsage$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScorePlayersUsage$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScorePlayersUsage$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScorePlayersUsage$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScorePlayersUsage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersUsage$new(
  player,
  team,
  position,
  total,
  quarter1,
  quarter2,
  quarter3,
  quarter4,
  rushing,
  passing
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersUsage$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersUsage$fromJSON(BoxScorePlayersUsageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersUsage$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersUsage$fromJSONString(BoxScorePlayersUsageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScorePlayersUsage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
