% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScoreTeams.r
\name{BoxScoreTeams}
\alias{BoxScoreTeams}
\title{BoxScoreTeams Class}
\description{
BoxScoreTeams Class

BoxScoreTeams Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ppa}}{}

\item{\code{cumulativePpa}}{}

\item{\code{successRates}}{}

\item{\code{explosiveness}}{}

\item{\code{rushing}}{}

\item{\code{havoc}}{}

\item{\code{scoringOpportunities}}{}

\item{\code{fieldPosition}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScoreTeams$new()}}
\item \href{#method-toJSON}{\code{BoxScoreTeams$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScoreTeams$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScoreTeams$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScoreTeams$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScoreTeams$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeams$new(
  ppa,
  cumulativePpa,
  successRates,
  explosiveness,
  rushing,
  havoc,
  scoringOpportunities,
  fieldPosition
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeams$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeams$fromJSON(BoxScoreTeamsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeams$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeams$fromJSONString(BoxScoreTeamsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeams$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
