% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScoreTeamsFieldPosition.r
\name{BoxScoreTeamsFieldPosition}
\alias{BoxScoreTeamsFieldPosition}
\title{BoxScoreTeamsFieldPosition Class}
\description{
BoxScoreTeamsFieldPosition Class

BoxScoreTeamsFieldPosition Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{averageStart}}{}

\item{\code{averageStartingPredictedPoints}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScoreTeamsFieldPosition$new()}}
\item \href{#method-toJSON}{\code{BoxScoreTeamsFieldPosition$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScoreTeamsFieldPosition$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScoreTeamsFieldPosition$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScoreTeamsFieldPosition$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScoreTeamsFieldPosition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsFieldPosition$new(
  team,
  averageStart,
  averageStartingPredictedPoints
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsFieldPosition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsFieldPosition$fromJSON(BoxScoreTeamsFieldPositionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsFieldPosition$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsFieldPosition$fromJSONString(BoxScoreTeamsFieldPositionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsFieldPosition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
