% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScoreTeamsHavoc.r
\name{BoxScoreTeamsHavoc}
\alias{BoxScoreTeamsHavoc}
\title{BoxScoreTeamsHavoc Class}
\description{
BoxScoreTeamsHavoc Class

BoxScoreTeamsHavoc Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{total}}{}

\item{\code{frontSeven}}{}

\item{\code{db}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScoreTeamsHavoc$new()}}
\item \href{#method-toJSON}{\code{BoxScoreTeamsHavoc$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScoreTeamsHavoc$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScoreTeamsHavoc$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScoreTeamsHavoc$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScoreTeamsHavoc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsHavoc$new(team, total, frontSeven, db)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsHavoc$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsHavoc$fromJSON(BoxScoreTeamsHavocJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsHavoc$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsHavoc$fromJSONString(BoxScoreTeamsHavocJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsHavoc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
