% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScoreTeamsPpa.r
\name{BoxScoreTeamsPpa}
\alias{BoxScoreTeamsPpa}
\title{BoxScoreTeamsPpa Class}
\description{
BoxScoreTeamsPpa Class

BoxScoreTeamsPpa Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{plays}}{}

\item{\code{overall}}{}

\item{\code{passing}}{}

\item{\code{rushing}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScoreTeamsPpa$new()}}
\item \href{#method-toJSON}{\code{BoxScoreTeamsPpa$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScoreTeamsPpa$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScoreTeamsPpa$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScoreTeamsPpa$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScoreTeamsPpa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsPpa$new(team, plays, overall, passing, rushing)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsPpa$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsPpa$fromJSON(BoxScoreTeamsPpaJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsPpa$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsPpa$fromJSONString(BoxScoreTeamsPpaJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsPpa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
