% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScoreTeamsRushing.r
\name{BoxScoreTeamsRushing}
\alias{BoxScoreTeamsRushing}
\title{BoxScoreTeamsRushing Class}
\description{
BoxScoreTeamsRushing Class

BoxScoreTeamsRushing Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{powerSuccess}}{}

\item{\code{stuffRate}}{}

\item{\code{lineYards}}{}

\item{\code{lineYardsAverage}}{}

\item{\code{secondLevelYards}}{}

\item{\code{secondLevelYardsAverage}}{}

\item{\code{openFieldYards}}{}

\item{\code{openFieldYardsAverage}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScoreTeamsRushing$new()}}
\item \href{#method-toJSON}{\code{BoxScoreTeamsRushing$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScoreTeamsRushing$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScoreTeamsRushing$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScoreTeamsRushing$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScoreTeamsRushing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsRushing$new(
  team,
  powerSuccess,
  stuffRate,
  lineYards,
  lineYardsAverage,
  secondLevelYards,
  secondLevelYardsAverage,
  openFieldYards,
  openFieldYardsAverage
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsRushing$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsRushing$fromJSON(BoxScoreTeamsRushingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsRushing$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsRushing$fromJSONString(BoxScoreTeamsRushingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsRushing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
