% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScoreTeamsScoringOpportunities.r
\name{BoxScoreTeamsScoringOpportunities}
\alias{BoxScoreTeamsScoringOpportunities}
\title{BoxScoreTeamsScoringOpportunities Class}
\description{
BoxScoreTeamsScoringOpportunities Class

BoxScoreTeamsScoringOpportunities Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{opportunities}}{}

\item{\code{points}}{}

\item{\code{pointsPerOpportunity}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScoreTeamsScoringOpportunities$new()}}
\item \href{#method-toJSON}{\code{BoxScoreTeamsScoringOpportunities$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScoreTeamsScoringOpportunities$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScoreTeamsScoringOpportunities$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScoreTeamsScoringOpportunities$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScoreTeamsScoringOpportunities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsScoringOpportunities$new(
  team,
  opportunities,
  points,
  pointsPerOpportunity
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsScoringOpportunities$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsScoringOpportunities$fromJSON(
  BoxScoreTeamsScoringOpportunitiesJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsScoringOpportunities$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsScoringOpportunities$fromJSONString(
  BoxScoreTeamsScoringOpportunitiesJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsScoringOpportunities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
