% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxScoreTeamsSuccessRates.r
\name{BoxScoreTeamsSuccessRates}
\alias{BoxScoreTeamsSuccessRates}
\title{BoxScoreTeamsSuccessRates Class}
\description{
BoxScoreTeamsSuccessRates Class

BoxScoreTeamsSuccessRates Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{overall}}{}

\item{\code{standardDowns}}{}

\item{\code{passingDowns}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BoxScoreTeamsSuccessRates$new()}}
\item \href{#method-toJSON}{\code{BoxScoreTeamsSuccessRates$toJSON()}}
\item \href{#method-fromJSON}{\code{BoxScoreTeamsSuccessRates$fromJSON()}}
\item \href{#method-toJSONString}{\code{BoxScoreTeamsSuccessRates$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BoxScoreTeamsSuccessRates$fromJSONString()}}
\item \href{#method-clone}{\code{BoxScoreTeamsSuccessRates$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsSuccessRates$new(team, overall, standardDowns, passingDowns)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsSuccessRates$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsSuccessRates$fromJSON(BoxScoreTeamsSuccessRatesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsSuccessRates$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsSuccessRates$fromJSONString(BoxScoreTeamsSuccessRatesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BoxScoreTeamsSuccessRates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
