% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoachesApi.r
\name{CoachesApi}
\alias{CoachesApi}
\title{Coaches operations}
\description{
swagger.Coaches
}
\section{Methods}{

\describe{

get_coaches Coaching records and history

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CoachesApi$new()}}
\item \href{#method-get_coaches}{\code{CoachesApi$get_coaches()}}
\item \href{#method-clone}{\code{CoachesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoachesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_coaches"></a>}}
\if{latex}{\out{\hypertarget{method-get_coaches}{}}}
\subsection{Method \code{get_coaches()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoachesApi$get_coaches(
  first_name,
  last_name,
  team,
  year,
  min_year,
  max_year,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoachesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
