% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConferenceSPRating.r
\name{ConferenceSPRating}
\alias{ConferenceSPRating}
\title{ConferenceSPRating Class}
\description{
ConferenceSPRating Class

ConferenceSPRating Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{year}}{}

\item{\code{conference}}{}

\item{\code{rating}}{}

\item{\code{secondOrderWins}}{}

\item{\code{sos}}{}

\item{\code{offense}}{}

\item{\code{defense}}{}

\item{\code{specialTeams}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConferenceSPRating$new()}}
\item \href{#method-toJSON}{\code{ConferenceSPRating$toJSON()}}
\item \href{#method-fromJSON}{\code{ConferenceSPRating$fromJSON()}}
\item \href{#method-toJSONString}{\code{ConferenceSPRating$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ConferenceSPRating$fromJSONString()}}
\item \href{#method-clone}{\code{ConferenceSPRating$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRating$new(
  year,
  conference,
  rating,
  secondOrderWins,
  sos,
  offense,
  defense,
  specialTeams
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRating$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRating$fromJSON(ConferenceSPRatingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRating$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRating$fromJSONString(ConferenceSPRatingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRating$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
