% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConferenceSPRatingDefense.r
\name{ConferenceSPRatingDefense}
\alias{ConferenceSPRatingDefense}
\title{ConferenceSPRatingDefense Class}
\description{
ConferenceSPRatingDefense Class

ConferenceSPRatingDefense Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rating}}{}

\item{\code{success}}{}

\item{\code{explosiveness}}{}

\item{\code{rushing}}{}

\item{\code{pasing}}{}

\item{\code{standardDowns}}{}

\item{\code{passingDowns}}{}

\item{\code{havoc}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConferenceSPRatingDefense$new()}}
\item \href{#method-toJSON}{\code{ConferenceSPRatingDefense$toJSON()}}
\item \href{#method-fromJSON}{\code{ConferenceSPRatingDefense$fromJSON()}}
\item \href{#method-toJSONString}{\code{ConferenceSPRatingDefense$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ConferenceSPRatingDefense$fromJSONString()}}
\item \href{#method-clone}{\code{ConferenceSPRatingDefense$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRatingDefense$new(
  rating,
  success,
  explosiveness,
  rushing,
  pasing,
  standardDowns,
  passingDowns,
  havoc
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRatingDefense$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRatingDefense$fromJSON(ConferenceSPRatingDefenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRatingDefense$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRatingDefense$fromJSONString(ConferenceSPRatingDefenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferenceSPRatingDefense$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
