% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drive.r
\name{Drive}
\alias{Drive}
\title{Drive Class}
\description{
Drive Class

Drive Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{offense}}{}

\item{\code{offense_conference}}{}

\item{\code{defense}}{}

\item{\code{defense_conference}}{}

\item{\code{game_id}}{}

\item{\code{id}}{}

\item{\code{drive_number}}{}

\item{\code{scoring}}{}

\item{\code{start_period}}{}

\item{\code{start_yardline}}{}

\item{\code{start_yards_to_goal}}{}

\item{\code{start_time}}{}

\item{\code{end_period}}{}

\item{\code{end_yardline}}{}

\item{\code{end_yards_to_goal}}{}

\item{\code{end_time}}{}

\item{\code{plays}}{}

\item{\code{yards}}{}

\item{\code{drive_result}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Drive$new()}}
\item \href{#method-toJSON}{\code{Drive$toJSON()}}
\item \href{#method-fromJSON}{\code{Drive$fromJSON()}}
\item \href{#method-toJSONString}{\code{Drive$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Drive$fromJSONString()}}
\item \href{#method-clone}{\code{Drive$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Drive$new(
  offense,
  offense_conference,
  defense,
  defense_conference,
  game_id,
  id,
  drive_number,
  scoring,
  start_period,
  start_yardline,
  start_yards_to_goal,
  start_time,
  end_period,
  end_yardline,
  end_yards_to_goal,
  end_time,
  plays,
  yards,
  drive_result
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Drive$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Drive$fromJSON(DriveJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Drive$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Drive$fromJSONString(DriveJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Drive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
