% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrivesApi.r
\name{DrivesApi}
\alias{DrivesApi}
\title{Drives operations}
\description{
swagger.Drives
}
\section{Methods}{

\describe{

get_drvies Drive data and results

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DrivesApi$new()}}
\item \href{#method-get_drvies}{\code{DrivesApi$get_drvies()}}
\item \href{#method-clone}{\code{DrivesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrivesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_drvies"></a>}}
\if{latex}{\out{\hypertarget{method-get_drvies}{}}}
\subsection{Method \code{get_drvies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrivesApi$get_drvies(
  year,
  season_type,
  week,
  team,
  offense,
  defense,
  conference,
  offense_conference,
  defense_conference,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrivesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
