% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game.r
\name{Game}
\alias{Game}
\title{Game Class}
\description{
Game Class

Game Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{season}}{}

\item{\code{week}}{}

\item{\code{season_type}}{}

\item{\code{start_date}}{}

\item{\code{start_time_tbd}}{}

\item{\code{neutral_site}}{}

\item{\code{conference_game}}{}

\item{\code{attendance}}{}

\item{\code{venue_id}}{}

\item{\code{venue}}{}

\item{\code{home_id}}{}

\item{\code{home_team}}{}

\item{\code{home_conference}}{}

\item{\code{home_points}}{}

\item{\code{home_line_scores}}{}

\item{\code{home_post_win_prob}}{}

\item{\code{away_id}}{}

\item{\code{away_team}}{}

\item{\code{away_conference}}{}

\item{\code{away_points}}{}

\item{\code{away_line_scores}}{}

\item{\code{away_post_win_prob}}{}

\item{\code{excitement_index}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Game$new()}}
\item \href{#method-toJSON}{\code{Game$toJSON()}}
\item \href{#method-fromJSON}{\code{Game$fromJSON()}}
\item \href{#method-toJSONString}{\code{Game$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Game$fromJSONString()}}
\item \href{#method-clone}{\code{Game$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$new(
  id,
  season,
  week,
  season_type,
  start_date,
  start_time_tbd,
  neutral_site,
  conference_game,
  attendance,
  venue_id,
  venue,
  home_id,
  home_team,
  home_conference,
  home_points,
  home_line_scores,
  home_post_win_prob,
  away_id,
  away_team,
  away_conference,
  away_points,
  away_line_scores,
  away_post_win_prob,
  excitement_index
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$fromJSON(GameJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$fromJSONString(GameJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
