% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameMedia.r
\name{GameMedia}
\alias{GameMedia}
\title{GameMedia Class}
\description{
GameMedia Class

GameMedia Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{season}}{}

\item{\code{week}}{}

\item{\code{seasonType}}{}

\item{\code{startTime}}{}

\item{\code{isStartTimeTBD}}{}

\item{\code{homeTeam}}{}

\item{\code{homeConference}}{}

\item{\code{awayTeam}}{}

\item{\code{awayConference}}{}

\item{\code{mediaType}}{}

\item{\code{outlet}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameMedia$new()}}
\item \href{#method-toJSON}{\code{GameMedia$toJSON()}}
\item \href{#method-fromJSON}{\code{GameMedia$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameMedia$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameMedia$fromJSONString()}}
\item \href{#method-clone}{\code{GameMedia$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMedia$new(
  id,
  season,
  week,
  seasonType,
  startTime,
  isStartTimeTBD,
  homeTeam,
  homeConference,
  awayTeam,
  awayConference,
  mediaType,
  outlet
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMedia$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMedia$fromJSON(GameMediaJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMedia$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMedia$fromJSONString(GameMediaJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMedia$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
