% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GamesApi.r
\name{GamesApi}
\alias{GamesApi}
\title{Games operations}
\description{
swagger.Games
}
\section{Methods}{

\describe{

get_advanced_box_score Advanced box scores


get_calendar Season calendar


get_game_media Game media information and schedules


get_games Games and results


get_player_game_stats Player game stats


get_team_game_stats Team game stats


get_team_records Team records

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GamesApi$new()}}
\item \href{#method-get_advanced_box_score}{\code{GamesApi$get_advanced_box_score()}}
\item \href{#method-get_calendar}{\code{GamesApi$get_calendar()}}
\item \href{#method-get_game_media}{\code{GamesApi$get_game_media()}}
\item \href{#method-get_games}{\code{GamesApi$get_games()}}
\item \href{#method-get_player_game_stats}{\code{GamesApi$get_player_game_stats()}}
\item \href{#method-get_team_game_stats}{\code{GamesApi$get_team_game_stats()}}
\item \href{#method-get_team_records}{\code{GamesApi$get_team_records()}}
\item \href{#method-clone}{\code{GamesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_advanced_box_score"></a>}}
\if{latex}{\out{\hypertarget{method-get_advanced_box_score}{}}}
\subsection{Method \code{get_advanced_box_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_advanced_box_score(game_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_calendar"></a>}}
\if{latex}{\out{\hypertarget{method-get_calendar}{}}}
\subsection{Method \code{get_calendar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_calendar(year, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_game_media"></a>}}
\if{latex}{\out{\hypertarget{method-get_game_media}{}}}
\subsection{Method \code{get_game_media()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_game_media(
  year,
  week,
  season_type,
  team,
  conference,
  media_type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_games"></a>}}
\if{latex}{\out{\hypertarget{method-get_games}{}}}
\subsection{Method \code{get_games()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_games(
  year,
  week,
  season_type,
  team,
  home,
  away,
  conference,
  id,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_player_game_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_player_game_stats}{}}}
\subsection{Method \code{get_player_game_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_player_game_stats(
  year,
  week,
  season_type,
  team,
  conference,
  category,
  game_id,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team_game_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_team_game_stats}{}}}
\subsection{Method \code{get_team_game_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_team_game_stats(
  year,
  week,
  season_type,
  team,
  conference,
  game_id,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team_records"></a>}}
\if{latex}{\out{\hypertarget{method-get_team_records}{}}}
\subsection{Method \code{get_team_records()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_team_records(year, team, conference, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
