% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsApi.r
\name{MetricsApi}
\alias{MetricsApi}
\title{Metrics operations}
\description{
swagger.Metrics
}
\section{Methods}{

\describe{

get_game_ppa Team Predicated Points Added (PPA/EPA) by game


get_player_game_ppa Player Predicated Points Added (PPA/EPA) broken down by game


get_player_season_ppa Player Predicated Points Added (PPA/EPA) broken down by season


get_predicted_points Predicted Points (i.e. Expected Points or EP)


get_pregame_win_probabilities Pregame win probability data


get_team_ppa Predicted Points Added (PPA/EPA) data by team


get_win_probability_data Win probability chart data

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetricsApi$new()}}
\item \href{#method-get_game_ppa}{\code{MetricsApi$get_game_ppa()}}
\item \href{#method-get_player_game_ppa}{\code{MetricsApi$get_player_game_ppa()}}
\item \href{#method-get_player_season_ppa}{\code{MetricsApi$get_player_season_ppa()}}
\item \href{#method-get_predicted_points}{\code{MetricsApi$get_predicted_points()}}
\item \href{#method-get_pregame_win_probabilities}{\code{MetricsApi$get_pregame_win_probabilities()}}
\item \href{#method-get_team_ppa}{\code{MetricsApi$get_team_ppa()}}
\item \href{#method-get_win_probability_data}{\code{MetricsApi$get_win_probability_data()}}
\item \href{#method-clone}{\code{MetricsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_game_ppa"></a>}}
\if{latex}{\out{\hypertarget{method-get_game_ppa}{}}}
\subsection{Method \code{get_game_ppa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$get_game_ppa(
  year,
  week,
  team,
  conference,
  exclude_garbage_time,
  season_type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_player_game_ppa"></a>}}
\if{latex}{\out{\hypertarget{method-get_player_game_ppa}{}}}
\subsection{Method \code{get_player_game_ppa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$get_player_game_ppa(
  year,
  week,
  team,
  position,
  player_id,
  threshold,
  exclude_garbage_time,
  season_type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_player_season_ppa"></a>}}
\if{latex}{\out{\hypertarget{method-get_player_season_ppa}{}}}
\subsection{Method \code{get_player_season_ppa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$get_player_season_ppa(
  year,
  team,
  conference,
  position,
  player_id,
  threshold,
  exclude_garbage_time,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_predicted_points"></a>}}
\if{latex}{\out{\hypertarget{method-get_predicted_points}{}}}
\subsection{Method \code{get_predicted_points()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$get_predicted_points(down, distance, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pregame_win_probabilities"></a>}}
\if{latex}{\out{\hypertarget{method-get_pregame_win_probabilities}{}}}
\subsection{Method \code{get_pregame_win_probabilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$get_pregame_win_probabilities(year, week, team, season_type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team_ppa"></a>}}
\if{latex}{\out{\hypertarget{method-get_team_ppa}{}}}
\subsection{Method \code{get_team_ppa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$get_team_ppa(year, team, conference, exclude_garbage_time, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_win_probability_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_win_probability_data}{}}}
\subsection{Method \code{get_win_probability_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$get_win_probability_data(game_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetricsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
