% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerGamePPAAveragePPA.r
\name{PlayerGamePPAAveragePPA}
\alias{PlayerGamePPAAveragePPA}
\title{PlayerGamePPAAveragePPA Class}
\description{
PlayerGamePPAAveragePPA Class

PlayerGamePPAAveragePPA Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{all}}{}

\item{\code{pass}}{}

\item{\code{rush}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayerGamePPAAveragePPA$new()}}
\item \href{#method-toJSON}{\code{PlayerGamePPAAveragePPA$toJSON()}}
\item \href{#method-fromJSON}{\code{PlayerGamePPAAveragePPA$fromJSON()}}
\item \href{#method-toJSONString}{\code{PlayerGamePPAAveragePPA$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PlayerGamePPAAveragePPA$fromJSONString()}}
\item \href{#method-clone}{\code{PlayerGamePPAAveragePPA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerGamePPAAveragePPA$new(all, pass, rush)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerGamePPAAveragePPA$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerGamePPAAveragePPA$fromJSON(PlayerGamePPAAveragePPAJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerGamePPAAveragePPA$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerGamePPAAveragePPA$fromJSONString(PlayerGamePPAAveragePPAJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerGamePPAAveragePPA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
