% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerSearchResult.r
\name{PlayerSearchResult}
\alias{PlayerSearchResult}
\title{PlayerSearchResult Class}
\description{
PlayerSearchResult Class

PlayerSearchResult Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{team}}{}

\item{\code{name}}{}

\item{\code{firstName}}{}

\item{\code{lastName}}{}

\item{\code{weight}}{}

\item{\code{height}}{}

\item{\code{jersey}}{}

\item{\code{position}}{}

\item{\code{hometown}}{}

\item{\code{teamColor}}{}

\item{\code{teamColorSecondary}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayerSearchResult$new()}}
\item \href{#method-toJSON}{\code{PlayerSearchResult$toJSON()}}
\item \href{#method-fromJSON}{\code{PlayerSearchResult$fromJSON()}}
\item \href{#method-toJSONString}{\code{PlayerSearchResult$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PlayerSearchResult$fromJSONString()}}
\item \href{#method-clone}{\code{PlayerSearchResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSearchResult$new(
  id,
  team,
  name,
  firstName,
  lastName,
  weight,
  height,
  jersey,
  position,
  hometown,
  teamColor,
  teamColorSecondary
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSearchResult$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSearchResult$fromJSON(PlayerSearchResultJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSearchResult$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSearchResult$fromJSONString(PlayerSearchResultJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSearchResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
