% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerSeasonPPA.r
\name{PlayerSeasonPPA}
\alias{PlayerSeasonPPA}
\title{PlayerSeasonPPA Class}
\description{
PlayerSeasonPPA Class

PlayerSeasonPPA Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{season}}{}

\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{position}}{}

\item{\code{team}}{}

\item{\code{conference}}{}

\item{\code{averagePPA}}{}

\item{\code{totalPPA}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayerSeasonPPA$new()}}
\item \href{#method-toJSON}{\code{PlayerSeasonPPA$toJSON()}}
\item \href{#method-fromJSON}{\code{PlayerSeasonPPA$fromJSON()}}
\item \href{#method-toJSONString}{\code{PlayerSeasonPPA$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PlayerSeasonPPA$fromJSONString()}}
\item \href{#method-clone}{\code{PlayerSeasonPPA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPA$new(
  season,
  id,
  name,
  position,
  team,
  conference,
  averagePPA,
  totalPPA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPA$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPA$fromJSON(PlayerSeasonPPAJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPA$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPA$fromJSONString(PlayerSeasonPPAJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
