% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerSeasonPPAAveragePPA.r
\name{PlayerSeasonPPAAveragePPA}
\alias{PlayerSeasonPPAAveragePPA}
\title{PlayerSeasonPPAAveragePPA Class}
\description{
PlayerSeasonPPAAveragePPA Class

PlayerSeasonPPAAveragePPA Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{all}}{}

\item{\code{pass}}{}

\item{\code{rush}}{}

\item{\code{firstDown}}{}

\item{\code{secondDown}}{}

\item{\code{thirdDown}}{}

\item{\code{standardDowns}}{}

\item{\code{passingDowns}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayerSeasonPPAAveragePPA$new()}}
\item \href{#method-toJSON}{\code{PlayerSeasonPPAAveragePPA$toJSON()}}
\item \href{#method-fromJSON}{\code{PlayerSeasonPPAAveragePPA$fromJSON()}}
\item \href{#method-toJSONString}{\code{PlayerSeasonPPAAveragePPA$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PlayerSeasonPPAAveragePPA$fromJSONString()}}
\item \href{#method-clone}{\code{PlayerSeasonPPAAveragePPA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPAAveragePPA$new(
  all,
  pass,
  rush,
  firstDown,
  secondDown,
  thirdDown,
  standardDowns,
  passingDowns
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPAAveragePPA$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPAAveragePPA$fromJSON(PlayerSeasonPPAAveragePPAJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPAAveragePPA$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPAAveragePPA$fromJSONString(PlayerSeasonPPAAveragePPAJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerSeasonPPAAveragePPA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
