% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerUsageUsage.r
\name{PlayerUsageUsage}
\alias{PlayerUsageUsage}
\title{PlayerUsageUsage Class}
\description{
PlayerUsageUsage Class

PlayerUsageUsage Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{overall}}{}

\item{\code{pass}}{}

\item{\code{rush}}{}

\item{\code{firstDown}}{}

\item{\code{secondDown}}{}

\item{\code{thirdDown}}{}

\item{\code{standardDowns}}{}

\item{\code{passingDowns}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayerUsageUsage$new()}}
\item \href{#method-toJSON}{\code{PlayerUsageUsage$toJSON()}}
\item \href{#method-fromJSON}{\code{PlayerUsageUsage$fromJSON()}}
\item \href{#method-toJSONString}{\code{PlayerUsageUsage$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PlayerUsageUsage$fromJSONString()}}
\item \href{#method-clone}{\code{PlayerUsageUsage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerUsageUsage$new(
  overall,
  pass,
  rush,
  firstDown,
  secondDown,
  thirdDown,
  standardDowns,
  passingDowns
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerUsageUsage$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerUsageUsage$fromJSON(PlayerUsageUsageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerUsageUsage$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerUsageUsage$fromJSONString(PlayerUsageUsageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerUsageUsage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
