% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaysApi.r
\name{PlaysApi}
\alias{PlaysApi}
\title{Plays operations}
\description{
swagger.Plays
}
\section{Methods}{

\describe{

get_play_stat_types Types of player play stats


get_play_stats Play stats by play


get_play_types Play types


get_plays Play by play data

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlaysApi$new()}}
\item \href{#method-get_play_stat_types}{\code{PlaysApi$get_play_stat_types()}}
\item \href{#method-get_play_stats}{\code{PlaysApi$get_play_stats()}}
\item \href{#method-get_play_types}{\code{PlaysApi$get_play_types()}}
\item \href{#method-get_plays}{\code{PlaysApi$get_plays()}}
\item \href{#method-clone}{\code{PlaysApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaysApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_play_stat_types"></a>}}
\if{latex}{\out{\hypertarget{method-get_play_stat_types}{}}}
\subsection{Method \code{get_play_stat_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaysApi$get_play_stat_types(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_play_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_play_stats}{}}}
\subsection{Method \code{get_play_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaysApi$get_play_stats(
  year,
  week,
  team,
  game_id,
  athlete_id,
  stat_type_id,
  season_type,
  conference,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_play_types"></a>}}
\if{latex}{\out{\hypertarget{method-get_play_types}{}}}
\subsection{Method \code{get_play_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaysApi$get_play_types(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_plays"></a>}}
\if{latex}{\out{\hypertarget{method-get_plays}{}}}
\subsection{Method \code{get_plays()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaysApi$get_plays(
  year,
  week,
  season_type,
  team,
  offense,
  defense,
  conference,
  offense_conference,
  defense_conference,
  play_type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlaysApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
