% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayWP.r
\name{PlayWP}
\alias{PlayWP}
\title{PlayWP Class}
\description{
PlayWP Class

PlayWP Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gamesId}}{}

\item{\code{playId}}{}

\item{\code{playText}}{}

\item{\code{homeId}}{}

\item{\code{home}}{}

\item{\code{awayId}}{}

\item{\code{away}}{}

\item{\code{spread}}{}

\item{\code{homeBall}}{}

\item{\code{homeScore}}{}

\item{\code{awayScore}}{}

\item{\code{timeRemaining}}{}

\item{\code{yardLine}}{}

\item{\code{down}}{}

\item{\code{distance}}{}

\item{\code{homeWinProb}}{}

\item{\code{playNumber}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayWP$new()}}
\item \href{#method-toJSON}{\code{PlayWP$toJSON()}}
\item \href{#method-fromJSON}{\code{PlayWP$fromJSON()}}
\item \href{#method-toJSONString}{\code{PlayWP$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PlayWP$fromJSONString()}}
\item \href{#method-clone}{\code{PlayWP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayWP$new(
  gamesId,
  playId,
  playText,
  homeId,
  home,
  awayId,
  away,
  spread,
  homeBall,
  homeScore,
  awayScore,
  timeRemaining,
  yardLine,
  down,
  distance,
  homeWinProb,
  playNumber
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayWP$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayWP$fromJSON(PlayWPJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayWP$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayWP$fromJSONString(PlayWPJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayWP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
