% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionGroupRecruitingRating.r
\name{PositionGroupRecruitingRating}
\alias{PositionGroupRecruitingRating}
\title{PositionGroupRecruitingRating Class}
\description{
PositionGroupRecruitingRating Class

PositionGroupRecruitingRating Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{conference}}{}

\item{\code{positiionGroup}}{}

\item{\code{averageRating}}{}

\item{\code{totalRating}}{}

\item{\code{commits}}{}

\item{\code{averageStars}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PositionGroupRecruitingRating$new()}}
\item \href{#method-toJSON}{\code{PositionGroupRecruitingRating$toJSON()}}
\item \href{#method-fromJSON}{\code{PositionGroupRecruitingRating$fromJSON()}}
\item \href{#method-toJSONString}{\code{PositionGroupRecruitingRating$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PositionGroupRecruitingRating$fromJSONString()}}
\item \href{#method-clone}{\code{PositionGroupRecruitingRating$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionGroupRecruitingRating$new(
  team,
  conference,
  positiionGroup,
  averageRating,
  totalRating,
  commits,
  averageStars
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionGroupRecruitingRating$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionGroupRecruitingRating$fromJSON(PositionGroupRecruitingRatingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionGroupRecruitingRating$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionGroupRecruitingRating$fromJSONString(PositionGroupRecruitingRatingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionGroupRecruitingRating$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
