% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatingsApi.r
\name{RatingsApi}
\alias{RatingsApi}
\title{Ratings operations}
\description{
swagger.Ratings
}
\section{Methods}{

\describe{

get_conference_sp_ratings Historical SP+ ratings by conference


get_sp_ratings Historical SP+ ratings


get_srs_ratings Historical SRS ratings

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RatingsApi$new()}}
\item \href{#method-get_conference_sp_ratings}{\code{RatingsApi$get_conference_sp_ratings()}}
\item \href{#method-get_sp_ratings}{\code{RatingsApi$get_sp_ratings()}}
\item \href{#method-get_srs_ratings}{\code{RatingsApi$get_srs_ratings()}}
\item \href{#method-clone}{\code{RatingsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_conference_sp_ratings"></a>}}
\if{latex}{\out{\hypertarget{method-get_conference_sp_ratings}{}}}
\subsection{Method \code{get_conference_sp_ratings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingsApi$get_conference_sp_ratings(year, conference, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sp_ratings"></a>}}
\if{latex}{\out{\hypertarget{method-get_sp_ratings}{}}}
\subsection{Method \code{get_sp_ratings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingsApi$get_sp_ratings(year, team, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_srs_ratings"></a>}}
\if{latex}{\out{\hypertarget{method-get_srs_ratings}{}}}
\subsection{Method \code{get_srs_ratings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingsApi$get_srs_ratings(year, team, conference, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatingsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
