% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruit.r
\name{Recruit}
\alias{Recruit}
\title{Recruit Class}
\description{
Recruit Class

Recruit Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{recruitType}}{}

\item{\code{year}}{}

\item{\code{ranking}}{}

\item{\code{name}}{}

\item{\code{school}}{}

\item{\code{committedTo}}{}

\item{\code{position}}{}

\item{\code{height}}{}

\item{\code{weight}}{}

\item{\code{stars}}{}

\item{\code{rating}}{}

\item{\code{city}}{}

\item{\code{stateProvince}}{}

\item{\code{country}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Recruit$new()}}
\item \href{#method-toJSON}{\code{Recruit$toJSON()}}
\item \href{#method-fromJSON}{\code{Recruit$fromJSON()}}
\item \href{#method-toJSONString}{\code{Recruit$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Recruit$fromJSONString()}}
\item \href{#method-clone}{\code{Recruit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruit$new(
  recruitType,
  year,
  ranking,
  name,
  school,
  committedTo,
  position,
  height,
  weight,
  stars,
  rating,
  city,
  stateProvince,
  country
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruit$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruit$fromJSON(RecruitJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruit$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruit$fromJSONString(RecruitJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recruit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
