% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitingApi.r
\name{RecruitingApi}
\alias{RecruitingApi}
\title{Recruiting operations}
\description{
swagger.Recruiting
}
\section{Methods}{

\describe{

get_recruiting_groups Recruit position group ratings


get_recruiting_players Player recruiting ratings and rankings


get_recruiting_teams Team recruiting rankings and ratings

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RecruitingApi$new()}}
\item \href{#method-get_recruiting_groups}{\code{RecruitingApi$get_recruiting_groups()}}
\item \href{#method-get_recruiting_players}{\code{RecruitingApi$get_recruiting_players()}}
\item \href{#method-get_recruiting_teams}{\code{RecruitingApi$get_recruiting_teams()}}
\item \href{#method-clone}{\code{RecruitingApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecruitingApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_recruiting_groups"></a>}}
\if{latex}{\out{\hypertarget{method-get_recruiting_groups}{}}}
\subsection{Method \code{get_recruiting_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecruitingApi$get_recruiting_groups(
  start_year,
  end_year,
  team,
  conference,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_recruiting_players"></a>}}
\if{latex}{\out{\hypertarget{method-get_recruiting_players}{}}}
\subsection{Method \code{get_recruiting_players()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecruitingApi$get_recruiting_players(
  year,
  classification,
  position,
  state,
  team,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_recruiting_teams"></a>}}
\if{latex}{\out{\hypertarget{method-get_recruiting_teams}{}}}
\subsection{Method \code{get_recruiting_teams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecruitingApi$get_recruiting_teams(year, team, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecruitingApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
