% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturningProduction.r
\name{ReturningProduction}
\alias{ReturningProduction}
\title{ReturningProduction Class}
\description{
ReturningProduction Class

ReturningProduction Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{season}}{}

\item{\code{team}}{}

\item{\code{conference}}{}

\item{\code{totalPPA}}{}

\item{\code{totalPassingPPA}}{}

\item{\code{totalReceivingPPA}}{}

\item{\code{totalRushingPPA}}{}

\item{\code{percentPPA}}{}

\item{\code{percentPassingPPA}}{}

\item{\code{percentReceivingPPA}}{}

\item{\code{percentRushingPPA}}{}

\item{\code{usage}}{}

\item{\code{passingUsage}}{}

\item{\code{receivingUsage}}{}

\item{\code{rushingUsage}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReturningProduction$new()}}
\item \href{#method-toJSON}{\code{ReturningProduction$toJSON()}}
\item \href{#method-fromJSON}{\code{ReturningProduction$fromJSON()}}
\item \href{#method-toJSONString}{\code{ReturningProduction$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ReturningProduction$fromJSONString()}}
\item \href{#method-clone}{\code{ReturningProduction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReturningProduction$new(
  season,
  team,
  conference,
  totalPPA,
  totalPassingPPA,
  totalReceivingPPA,
  totalRushingPPA,
  percentPPA,
  percentPassingPPA,
  percentReceivingPPA,
  percentRushingPPA,
  usage,
  passingUsage,
  receivingUsage,
  rushingUsage
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReturningProduction$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReturningProduction$fromJSON(ReturningProductionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReturningProduction$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReturningProduction$fromJSONString(ReturningProductionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReturningProduction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
