% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsApi.r
\name{StatsApi}
\alias{StatsApi}
\title{Stats operations}
\description{
swagger.Stats
}
\section{Methods}{

\describe{

get_advanced_team_game_stats Advanced team metrics by game


get_advanced_team_season_stats Advanced team metrics by season


get_stat_categories Team stat categories


get_team_season_stats Team statistics by season

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StatsApi$new()}}
\item \href{#method-get_advanced_team_game_stats}{\code{StatsApi$get_advanced_team_game_stats()}}
\item \href{#method-get_advanced_team_season_stats}{\code{StatsApi$get_advanced_team_season_stats()}}
\item \href{#method-get_stat_categories}{\code{StatsApi$get_stat_categories()}}
\item \href{#method-get_team_season_stats}{\code{StatsApi$get_team_season_stats()}}
\item \href{#method-clone}{\code{StatsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_advanced_team_game_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_advanced_team_game_stats}{}}}
\subsection{Method \code{get_advanced_team_game_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$get_advanced_team_game_stats(
  year,
  week,
  team,
  opponent,
  exclude_garbage_time,
  season_type,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_advanced_team_season_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_advanced_team_season_stats}{}}}
\subsection{Method \code{get_advanced_team_season_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$get_advanced_team_season_stats(
  year,
  team,
  exclude_garbage_time,
  start_week,
  end_week,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_stat_categories"></a>}}
\if{latex}{\out{\hypertarget{method-get_stat_categories}{}}}
\subsection{Method \code{get_stat_categories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$get_stat_categories(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team_season_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_team_season_stats}{}}}
\subsection{Method \code{get_team_season_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$get_team_season_stats(
  year,
  team,
  conference,
  start_week,
  end_week,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
