% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamMatchupGames.r
\name{TeamMatchupGames}
\alias{TeamMatchupGames}
\title{TeamMatchupGames Class}
\description{
TeamMatchupGames Class

TeamMatchupGames Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{season}}{}

\item{\code{week}}{}

\item{\code{season_type}}{}

\item{\code{date}}{}

\item{\code{neutralSite}}{}

\item{\code{venue}}{}

\item{\code{homeTeam}}{}

\item{\code{homeScore}}{}

\item{\code{awayTeam}}{}

\item{\code{awayScore}}{}

\item{\code{winner}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamMatchupGames$new()}}
\item \href{#method-toJSON}{\code{TeamMatchupGames$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamMatchupGames$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamMatchupGames$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamMatchupGames$fromJSONString()}}
\item \href{#method-clone}{\code{TeamMatchupGames$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamMatchupGames$new(
  season,
  week,
  season_type,
  date,
  neutralSite,
  venue,
  homeTeam,
  homeScore,
  awayTeam,
  awayScore,
  winner
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamMatchupGames$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamMatchupGames$fromJSON(TeamMatchupGamesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamMatchupGames$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamMatchupGames$fromJSONString(TeamMatchupGamesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamMatchupGames$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
