% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamPPAOffense.r
\name{TeamPPAOffense}
\alias{TeamPPAOffense}
\title{TeamPPAOffense Class}
\description{
TeamPPAOffense Class

TeamPPAOffense Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{overall}}{}

\item{\code{passing}}{}

\item{\code{rushing}}{}

\item{\code{firstDown}}{}

\item{\code{secondDown}}{}

\item{\code{thirdDown}}{}

\item{\code{cumulative}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamPPAOffense$new()}}
\item \href{#method-toJSON}{\code{TeamPPAOffense$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamPPAOffense$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamPPAOffense$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamPPAOffense$fromJSONString()}}
\item \href{#method-clone}{\code{TeamPPAOffense$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffense$new(
  overall,
  passing,
  rushing,
  firstDown,
  secondDown,
  thirdDown,
  cumulative
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffense$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffense$fromJSON(TeamPPAOffenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffense$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffense$fromJSONString(TeamPPAOffenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffense$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
