% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamPPAOffenseCumulative.r
\name{TeamPPAOffenseCumulative}
\alias{TeamPPAOffenseCumulative}
\title{TeamPPAOffenseCumulative Class}
\description{
TeamPPAOffenseCumulative Class

TeamPPAOffenseCumulative Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{total}}{}

\item{\code{passing}}{}

\item{\code{rushing}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamPPAOffenseCumulative$new()}}
\item \href{#method-toJSON}{\code{TeamPPAOffenseCumulative$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamPPAOffenseCumulative$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamPPAOffenseCumulative$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamPPAOffenseCumulative$fromJSONString()}}
\item \href{#method-clone}{\code{TeamPPAOffenseCumulative$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffenseCumulative$new(total, passing, rushing)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffenseCumulative$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffenseCumulative$fromJSON(TeamPPAOffenseCumulativeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffenseCumulative$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffenseCumulative$fromJSONString(TeamPPAOffenseCumulativeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamPPAOffenseCumulative$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
