% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamsApi.r
\name{TeamsApi}
\alias{TeamsApi}
\title{Teams operations}
\description{
swagger.Teams
}
\section{Methods}{

\describe{

get_fbs_teams FBS team list


get_roster Team rosters


get_talent Team talent composite rankings


get_team_matchup Team matchup history


get_teams Team information

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamsApi$new()}}
\item \href{#method-get_fbs_teams}{\code{TeamsApi$get_fbs_teams()}}
\item \href{#method-get_roster}{\code{TeamsApi$get_roster()}}
\item \href{#method-get_talent}{\code{TeamsApi$get_talent()}}
\item \href{#method-get_team_matchup}{\code{TeamsApi$get_team_matchup()}}
\item \href{#method-get_teams}{\code{TeamsApi$get_teams()}}
\item \href{#method-clone}{\code{TeamsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fbs_teams"></a>}}
\if{latex}{\out{\hypertarget{method-get_fbs_teams}{}}}
\subsection{Method \code{get_fbs_teams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_fbs_teams(year, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_roster"></a>}}
\if{latex}{\out{\hypertarget{method-get_roster}{}}}
\subsection{Method \code{get_roster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_roster(team, year, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_talent"></a>}}
\if{latex}{\out{\hypertarget{method-get_talent}{}}}
\subsection{Method \code{get_talent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_talent(year, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team_matchup"></a>}}
\if{latex}{\out{\hypertarget{method-get_team_matchup}{}}}
\subsection{Method \code{get_team_matchup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_team_matchup(team1, team2, min_year, max_year, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_teams"></a>}}
\if{latex}{\out{\hypertarget{method-get_teams}{}}}
\subsection{Method \code{get_teams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_teams(conference, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
