% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamSeason.r
\name{TeamSeason}
\alias{TeamSeason}
\title{TeamSeason Class}
\description{
TeamSeason Class

TeamSeason Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{school}}{}

\item{\code{year}}{}

\item{\code{games}}{}

\item{\code{wins}}{}

\item{\code{losses}}{}

\item{\code{ties}}{}

\item{\code{preseason_rank}}{}

\item{\code{postseason_rank}}{}

\item{\code{srs}}{}

\item{\code{sp_overall}}{}

\item{\code{sp_offense}}{}

\item{\code{sp_defense}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamSeason$new()}}
\item \href{#method-toJSON}{\code{TeamSeason$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamSeason$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamSeason$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamSeason$fromJSONString()}}
\item \href{#method-clone}{\code{TeamSeason$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSeason$new(
  school,
  year,
  games,
  wins,
  losses,
  ties,
  preseason_rank,
  postseason_rank,
  srs,
  sp_overall,
  sp_offense,
  sp_defense
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSeason$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSeason$fromJSON(TeamSeasonJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSeason$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSeason$fromJSONString(TeamSeasonJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSeason$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
