% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamSPRatingDefense.r
\name{TeamSPRatingDefense}
\alias{TeamSPRatingDefense}
\title{TeamSPRatingDefense Class}
\description{
TeamSPRatingDefense Class

TeamSPRatingDefense Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ranking}}{}

\item{\code{rating}}{}

\item{\code{success}}{}

\item{\code{explosiveness}}{}

\item{\code{rushing}}{}

\item{\code{pasing}}{}

\item{\code{standardDowns}}{}

\item{\code{passingDowns}}{}

\item{\code{havoc}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamSPRatingDefense$new()}}
\item \href{#method-toJSON}{\code{TeamSPRatingDefense$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamSPRatingDefense$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamSPRatingDefense$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamSPRatingDefense$fromJSONString()}}
\item \href{#method-clone}{\code{TeamSPRatingDefense$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingDefense$new(
  ranking,
  rating,
  success,
  explosiveness,
  rushing,
  pasing,
  standardDowns,
  passingDowns,
  havoc
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingDefense$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingDefense$fromJSON(TeamSPRatingDefenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingDefense$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingDefense$fromJSONString(TeamSPRatingDefenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingDefense$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
