% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamSPRatingOffense.r
\name{TeamSPRatingOffense}
\alias{TeamSPRatingOffense}
\title{TeamSPRatingOffense Class}
\description{
TeamSPRatingOffense Class

TeamSPRatingOffense Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ranking}}{}

\item{\code{rating}}{}

\item{\code{success}}{}

\item{\code{explosiveness}}{}

\item{\code{rushing}}{}

\item{\code{passing}}{}

\item{\code{standardDowns}}{}

\item{\code{passingDowns}}{}

\item{\code{runRate}}{}

\item{\code{pace}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamSPRatingOffense$new()}}
\item \href{#method-toJSON}{\code{TeamSPRatingOffense$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamSPRatingOffense$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamSPRatingOffense$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamSPRatingOffense$fromJSONString()}}
\item \href{#method-clone}{\code{TeamSPRatingOffense$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingOffense$new(
  ranking,
  rating,
  success,
  explosiveness,
  rushing,
  passing,
  standardDowns,
  passingDowns,
  runRate,
  pace
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingOffense$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingOffense$fromJSON(TeamSPRatingOffenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingOffense$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingOffense$fromJSONString(TeamSPRatingOffenseJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamSPRatingOffense$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
